冻结角色

+++ chapter-25-shiro-2/src/main/java/com/yucong/dto/role/RoleIdDTO.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class RoleIdDTO {
 
 	@NotNull(message="roleId必填")
-	public Integer roleId;
+	public Long roleId;
 }
